requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.3
crispy-bootstrap5==0.4
Django==3.2.5
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-js-asset==1.2.2
django-unused-media==0.2.2
idna==3.2
Pillow==8.3.1
psycopg2==2.9.1
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.6
